+++ src/ff7_data.h
@@ -221,9 +221,12 @@ void ff7_find_externals(struct ff7_game_obj* game_object)
 	ff7_externals.field_models_eye_to_model = get_relative_call(ff7_externals.field_load_models, 0xA79);
 	ff7_externals.field_load_animation = get_relative_call(ff7_externals.field_load_models, 0x8DF);
 	ff7_externals.field_load_model_eye_tex = (int (*)(ff7_model_eye_texture_data*,field_animation_data*))get_relative_call(ff7_externals.field_load_models, 0xB90);
-	ff7_externals.field_unload_model_eye_tex = (void (*)(void*))get_relative_call((uint32_t)ff7_externals.field_load_model_eye_tex, 0x1DE);
+	ff7_externals.field_load_model_tex = (p_hundred* (*)(int, int, char*, struc_3*, game_obj*))get_relative_call((uint32_t)ff7_externals.field_load_model_eye_tex, 0x9D);
+	ff7_externals.field_unload_model_tex = (void (*)(void*))get_relative_call((uint32_t)ff7_externals.field_load_model_eye_tex, 0x1DE);
+	ff7_externals.create_struc_3_info_sub_67455E = (void (*)(struc_3*))get_relative_call((uint32_t)ff7_externals.field_load_model_eye_tex, 0x12);
 	ff7_externals.load_animation = get_relative_call(ff7_externals.field_load_animation, 0x16D);
 	ff7_externals.destroy_animation = (void (*)(anim_header*))get_relative_call(ff7_externals.load_animation, 0x162);
+	ff7_externals.field_unk_909288 = (uint32_t*)get_absolute_value((uint32_t)ff7_externals.field_load_model_eye_tex, 0x1D);
 
 	ff7_externals.field_models_eye_blink_buffer = (ff7_model_eye_texture_data*)get_absolute_value(ff7_externals.field_load_models, 0xB8C);
 	ff7_externals.field_models_data = (DWORD*)get_absolute_value(ff7_externals.field_load_models, 0xE);
@@ -568,6 +571,7 @@ void ff7_find_externals(struct ff7_game_obj* game_object)
 	ff7_externals.field_evaluate_encounter_rate_60B2C6 = (void (*)())get_relative_call(ff7_externals.field_update_models_positions, 0x90F);
 	ff7_externals.field_animate_3d_models_6392BB = get_relative_call(field_main_loop, 0xF6);
 	ff7_externals.field_blink_3d_model_649B50 = (void(*)(field_animation_data*, field_model_blink_data*))get_relative_call(ff7_externals.field_animate_3d_models_6392BB, 0x8A7);
+	ff7_externals.field_sub_6A2736 = (int (*)(ff7_polygon_set*))get_relative_call((uint32_t)ff7_externals.field_blink_3d_model_649B50, 0xC4);
 	ff7_externals.field_model_blink_data_D000C8 = (field_model_blink_data*)get_absolute_value(ff7_externals.field_animate_3d_models_6392BB, 0x7E6);
 	ff7_externals.field_apply_kawai_op_64A070 = get_relative_call(ff7_externals.field_animate_3d_models_6392BB, 0x726);
 	ff7_externals.field_player_model_id = (short*)get_absolute_value(ff7_externals.field_update_models_positions, 0x45D);