+++ src/field.cpp
@@ -39,7 +39,9 @@ bool map_changing = false;
 int (*opcode_old_kawai)();
 int (*opcode_old_pc)();
 
+ff7_polygon_set* ff7_head = NULL;
 ff7_model_eye_texture_data ff7_eyes[9];
+ff7_model_mouth_data ff7_mouths[9];
 
 byte get_field_bank_value(int16_t bank)
 {
@@ -135,28 +137,62 @@ int opcode_kawai_eye_texture() {
 				}
 			}
 
+			// MOUTH
+			char* char_name = strtok(filename, "_");
+
+			_snprintf(directpath, sizeof(directpath), "%s/%s/flevel/mouth_%s_%d.TEX", basedir, direct_mode_path.c_str(), curr_model_name, mouth_index);
+			if (ff7_mouths[curr_eye_index].has_mouth = fileExists(directpath))
+				_snprintf(ff7_mouths[curr_eye_index].mouth_tex_filename, 1024, "mouth_%s_%d%s", curr_model_name, mouth_index, ext);
+			else if (curr_eye_index < 9)
+			{
+				if (mouth_index > 0 && (trace_all || trace_direct || trace_opcodes)) ffnx_trace("subcode[EYETX]: Custom mouth texture not found: %s\n", directpath);
+
+				_snprintf(directpath, sizeof(directpath), "%s/%s/flevel/%s_mouth_%d.TEX", basedir, direct_mode_path.c_str(), filename, mouth_index);
+				if (ff7_mouths[curr_eye_index].has_mouth = fileExists(directpath))
+					_snprintf(ff7_mouths[curr_eye_index].mouth_tex_filename, 1024, "%s_mouth_%d%s", char_name, mouth_index, ext);
+				else if (mouth_index > 0 && (trace_all || trace_direct || trace_opcodes)) ffnx_trace("subcode[EYETX]: Custom mouth texture not found: %s\n", directpath);
+			}
+
+			// Prepare mouth tex object
+			if (ff7_mouths[curr_eye_index].has_mouth)
+			{
+				struc_3 tex_mouth_info;
+				ff7_externals.create_struc_3_info_sub_67455E(&tex_mouth_info);
+				tex_mouth_info.base_directory = (uint32_t)ff7_externals.field_unk_909288;
+				tex_mouth_info.file_context.use_lgp = 1;
+				tex_mouth_info.file_context.lgp_num = 1;
+				tex_mouth_info.file_context.name_mangler = 0;
+				ff7_mouths[curr_eye_index].mouth_tex = ff7_externals.field_load_model_tex(0, 0, ff7_mouths[curr_eye_index].mouth_tex_filename, &tex_mouth_info, common_externals.get_game_object());
+			}
+			else
+			{
+				if (ff7_mouths[curr_eye_index].mouth_tex) ff7_externals.field_unload_model_tex(ff7_mouths[curr_eye_index].mouth_tex);
+				ff7_mouths[curr_eye_index].mouth_tex = NULL;
+			}
+			ff7_mouths[curr_eye_index].current_mouth_idx = mouth_index;
+
 			// Reload TEX data in memory
-			if(animation_data[curr_model_id].static_left_eye_tex) ff7_externals.field_unload_model_eye_tex(animation_data[curr_model_id].static_left_eye_tex);
-			if(animation_data[curr_model_id].static_right_eye_tex) ff7_externals.field_unload_model_eye_tex(animation_data[curr_model_id].static_right_eye_tex);
+			if(animation_data[curr_model_id].static_left_eye_tex) ff7_externals.field_unload_model_tex(animation_data[curr_model_id].static_left_eye_tex);
+			if(animation_data[curr_model_id].static_right_eye_tex) ff7_externals.field_unload_model_tex(animation_data[curr_model_id].static_right_eye_tex);
 			ff7_externals.field_load_model_eye_tex(&ff7_externals.field_models_eye_blink_buffer[curr_eye_index], &animation_data[curr_model_id]);
 
 			// Index is also treated as blink mode, if higher than 2 then "fake a closed eyes" in order to reload textures
 			if (left_eye_index <= 2 || right_eye_index <= 2)
 			{
-				ff7_externals.field_model_blink_data_D000C8->blink_left_eye_mode = 0;
-				ff7_externals.field_model_blink_data_D000C8->blink_right_eye_mode = 0;
-				ff7_externals.field_model_blink_data_D000C8->unknown = 0;
-				ff7_externals.field_model_blink_data_D000C8->model_id = curr_model_id;
-				ff7_externals.field_blink_3d_model_649B50(&animation_data[curr_model_id], ff7_externals.field_model_blink_data_D000C8);
+				ff7_externals.field_model_blink_data_D000C8->blink_left_eye_mode = left_eye_index;
+				ff7_externals.field_model_blink_data_D000C8->blink_right_eye_mode = right_eye_index;
 			}
 			else
 			{
 				ff7_externals.field_model_blink_data_D000C8->blink_left_eye_mode = 2;
 				ff7_externals.field_model_blink_data_D000C8->blink_right_eye_mode = 2;
-				ff7_externals.field_model_blink_data_D000C8->unknown = 0;
-				ff7_externals.field_model_blink_data_D000C8->model_id = curr_model_id;
-				ff7_externals.field_blink_3d_model_649B50(&animation_data[curr_model_id], ff7_externals.field_model_blink_data_D000C8);
 			}
+			ff7_externals.field_model_blink_data_D000C8->unknown = 0;
+			ff7_externals.field_model_blink_data_D000C8->model_id = curr_model_id;
+			ff7_externals.field_blink_3d_model_649B50(&animation_data[curr_model_id], ff7_externals.field_model_blink_data_D000C8);
+
+			// Required to force reload the mouth texture
+			ff7_head->per_group_hundreds = 1;
 		}
 	}
 
@@ -188,6 +224,23 @@ int field_calc_window_pos(int16_t WINDOW_ID, int16_t X, int16_t Y, int16_t W, in
 	return ff7_externals.sub_630C48(WINDOW_ID, X, ff7_field_center ? Y + 8 : Y, W, H);
 }
 
+int field_load_mouth(ff7_polygon_set *polygon_set)
+{
+	int ret = ff7_externals.field_sub_6A2736(polygon_set);
+
+	field_animation_data* animation_data = *ff7_externals.field_animation_data_ptr;
+	byte curr_model_id = ff7_externals.field_model_id_array[*ff7_externals.current_entity_id];
+	byte curr_eye_index = animation_data[curr_model_id].eye_texture_idx;
+	int idx = polygon_set->numgroups - 1;
+
+	if (polygon_set && ff7_mouths[curr_eye_index].mouth_tex)
+		polygon_set->hundred_data_group_array[idx] = ff7_mouths[curr_eye_index].mouth_tex;
+
+	ff7_head = polygon_set;
+
+	return ret;
+}
+
 void field_init()
 {
 	if (!ff8)
@@ -201,6 +254,9 @@ void field_init()
 
 		// Proxy the window calculation formula so we can offset windows vertically
 		replace_call_function(common_externals.execute_opcode_table[0x50] + 0x174, field_calc_window_pos);
+
+		// Proxy the function to easily obtain the current polygon set for mouth replacement
+		replace_call_function((uint32_t)ff7_externals.field_blink_3d_model_649B50 + 0xC4, field_load_mouth);
 
 		// ################################
 		// save static eyes names for later
@@ -222,6 +278,10 @@ void field_init()
 					ff7_externals.field_models_eye_blink_buffer[i].static_right_eye_filename = (char*)external_malloc(1024);
 					strcpy(ff7_externals.field_models_eye_blink_buffer[i].static_right_eye_filename, ff7_eyes[i].static_right_eye_filename);
 				}
+
+				ff7_mouths[i].current_mouth_idx = 0;
+				ff7_mouths[i].mouth_tex_filename = (char*)external_calloc(sizeof(char), sizeof(basedir) + 1024);
+				ff7_mouths[i].mouth_tex = NULL;
 			}
 		}
 	}